+++ tests/Console/AssetCommandHandlerTest.php
@@ -209,6 +209,22 @@ EOF;
         $this->assertSame(0, $this->handler->handleMap($args));
     }
 
+    public function testMapForce()
+    {
+        $this->assetManager->expects($this->once())
+            ->method('addAssetMapping')
+            ->willReturnCallback(function (AssetMapping $mapping, $flags) {
+                PHPUnit_Framework_Assert::assertSame('/app/public', $mapping->getGlob());
+                PHPUnit_Framework_Assert::assertSame('/', $mapping->getWebPath());
+                PHPUnit_Framework_Assert::assertSame(InstallTarget::DEFAULT_TARGET, $mapping->getTargetName());
+                PHPUnit_Framework_Assert::assertSame(AssetManager::NO_TARGET_CHECK, $flags);
+            });
+
+        $args = self::$mapCommand->parseArgs(new StringArgs('--force /app/public /'));
+
+        $this->assertSame(0, $this->handler->handleMap($args));
+    }
+
     public function testRemoveMapping()
     {
         $this->assetManager->expects($this->at(0))