+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Changelog
  * renamed binding type "puli/web-resource" to "puli/asset-mapping"
  * renamed `ResourceUrlGenerator` to `AssetUrlGenerator`
  * added NO_TARGET_CHECK flag to suppress sanity check in `AssetManager::addAssetMapping()`
+ * added --force option to "asset map" command
  
 * 1.0.0-beta (2015-03-19)